Add macOS/Metal GPU backend and port samples to cross-platform
New Metal backend (src/graphics/metal/) implementing DeviceBackend with
CAMetalLayer swapchain, MSL shader compilation, render/compute pipelines,
descriptor sets, and resource management. Vertex buffers at slots 16+ to
avoid uniform buffer conflicts.

New macOS window backend (src/uix/Window.mm) with NSWindow/PicoView/
CAMetalLayer, Retina backing scale support, and event forwarding to ImGui.
ImGui Metal+OSX backends (imgui_impl_metal/osx) integrated with proper
render pass descriptor handoff.

Cross-platform fixes: unified SwapchainInit with void* windowHandle,
unified shader source (single string, two entry points), replaced
platform-specific debug output with picoLogf, fixed Clang compilation
issues (Graph.h default initializer, StreamLayout size_t templates,
Model.cpp includes, ModelDraw.cpp narrowing casts).

Samples pico_01/02/03 run on macOS with correct Retina resolution,
ImGui mouse interaction, and point cloud rendering.

+++ src/document/Model.cpp
@@ -33,6 +33,8 @@
 
 #include <vector>
 #include <algorithm>
+#include <cfloat>
+#include <cstdint>
 
 #include <core/json/json.h>
 #include <core/Log.h>
@@ -159,7 +161,7 @@ std::vector<uint8_t> parseBinary(const uint64_t byteLength, const std::filesyste
     std::vector<uint8_t> bytes;
     if (std::filesystem::exists(glb_path)) {
         auto fileSize = std::filesystem::file_size(glb_path);
-        auto readSize = std::min(fileSize, byteLength);
+        auto readSize = std::min((uint64_t)fileSize, byteLength);
         
         std::ifstream  src(glb_path.c_str(), std::ios::binary | std::ios::in);
         if(!src.fail()) {
